Initial Upload
This program processes data from a pulsar data txt file and stores them in 5 parallel arrays.
Then, it creates a filtered txt file based on a sigma value threshold, and a dmBins file in which the largest sigma of each dm value is output.
Fix Later: dmBins function should not use a double variable loop counter